// packages/oos-dsls-ts/src/generated/ast.ts
export interface Accordion extends langium.AstNode {
    readonly $container: AccordionItem | Card | Grid | Row | Section | Stack | Tab | ViewDecl;
    readonly $type: 'Accordion';
    items: Array<AccordionItem>;
    mods: Array<LayoutMod>;
}

export const Accordion = {
    $type: 'Accordion',
    items: 'items',
    mods: 'mods'
} as const;

export function isAccordion(item: unknown): item is Accordion {
    return reflection.isInstance(item, Accordion.$type);
}

export interface AccordionItem extends langium.AstNode {
    readonly $container: Accordion;
    readonly $type: 'AccordionItem';
    body: Array<BodyElement>;
    caption: string;
    open: boolean;
}

export const AccordionItem = {
    $type: 'AccordionItem',
    body: 'body',
    caption: 'caption',
    open: 'open'
} as const;
// ...
export interface Bind extends langium.AstNode {
    readonly $container: ToolbarNew | ViewAction;
    readonly $type: 'Bind';
    source: FieldRef;
}

export const Bind = {
    $type: 'Bind',
    source: 'source'
} as const;

export function isBind(item: unknown): item is Bind {
    return reflection.isInstance(item, Bind.$type);
}

export type BodyElement = Accordion | Button | Card | Divider | Grid | IconWidget | LinkWidget | RichText | Row | Section | Separator | Stack | Table | Tabs | Widget;

export const BodyElement = {
    $type: 'BodyElement'
} as const;

export function isBodyElement(item: unknown): item is BodyElement {
    return reflection.isInstance(item, BodyElement.$type);
}

export type BoolLit = 'false' | 'true';

export function isBoolLit(item: unknown): item is BoolLit {
    return item === 'true' || item === 'false';
}

export interface Button extends langium.AstNode {
    readonly $container: AccordionItem | Card | Grid | Row | Section | Stack | Tab | ViewDecl;
    readonly $type: 'Button';
    actionRef?: string;
    caption: string;
    mods: Array<LayoutMod>;
}

export const Button = {
    $type: 'Button',
    actionRef: 'actionRef',
    caption: 'caption',
    mods: 'mods'
} as const;

export function isButton(item: unknown): item is Button {
    return reflection.isInstance(item, Button.$type);
}

export interface Card extends langium.AstNode {
    readonly $container: AccordionItem | Card | Grid | Row | Section | Stack | Tab | ViewDecl;
    readonly $type: 'Card';
    body: Array<BodyElement>;
    caption?: string;
    mods: Array<LayoutMod>;
}

export const Card = {
    $type: 'Card',
    body: 'body',
    caption: 'caption',
    mods: 'mods'
} as const;

export function isCard(item: unknown): item is Card {
    return reflection.isInstance(item, Card.$type);
}

export interface Cols extends langium.AstNode {
    readonly $container: Accordion | Button | Card | Grid | IconWidget | LinkWidget | Row | Section | Stack | Tabs | Widget;
    readonly $type: 'Cols';
    value: number;
}

export const Cols = {
    $type: 'Cols',
    value: 'value'
} as const;

export function isCols(item: unknown): item is Cols {
    return reflection.isInstance(item, Cols.$type);
}

export interface Column extends langium.AstNode {
    readonly $container: Table;
    readonly $type: 'Column';
    caption: string;
    field: FieldRef;
    mods: Array<ColumnModifier>;
}

export const Column = {
    $type: 'Column',
    caption: 'caption',
    field: 'field',
    mods: 'mods'
} as const;

export function isColumn(item: unknown): item is Column {
    return reflection.isInstance(item, Column.$type);
}

export interface ColumnFormat extends langium.AstNode {
    readonly $container: Column;
    readonly $type: 'ColumnFormat';
    fmt: FormatToken;
}

export const ColumnFormat = {
    $type: 'ColumnFormat',
    fmt: 'fmt'
} as const;

export function isColumnFormat(item: unknown): item is ColumnFormat {
    return reflection.isInstance(item, ColumnFormat.$type);
}

export type ColumnModifier = ColumnFormat | ColumnWidth;

export const ColumnModifier = {
    $type: 'ColumnModifier'
} as const;
// ...
export interface Expand extends langium.AstNode {
    readonly $container: Accordion | Button | Card | Grid | IconWidget | LinkWidget | Row | Section | Stack | Tabs | Widget;
    readonly $type: 'Expand';
}

export const Expand = {
    $type: 'Expand'
} as const;
// ...
export interface FieldRef extends langium.AstNode {
    readonly $container: Bind | Column | Table | Widget;
    readonly $type: 'FieldRef';
    domain: IdentName;
    field: IdentName;
}

export const FieldRef = {
    $type: 'FieldRef',
    domain: 'domain',
    field: 'field'
} as const;
// ...
export type FormatDetail = 'full' | 'long' | 'medium' | 'short' | number;


export type FormatKind = 'currency' | 'date' | 'datetime' | 'number' | 'percent' | 'time';

export function isFormatKind(item: unknown): item is FormatKind {
    return item === 'currency' || item === 'number' || item === 'percent' || item === 'datetime' || item === 'date' || item === 'time';
}

export interface FormatToken extends langium.AstNode {
    readonly $container: ColumnFormat | WidgetFormat;
    readonly $type: 'FormatToken';
    detail?: FormatDetail;
    kind: FormatKind;
}

export const FormatToken = {
    $type: 'FormatToken',
    detail: 'detail',
    kind: 'kind'
} as const;

export function isFormatToken(item: unknown): item is FormatToken {
    return reflection.isInstance(item, FormatToken.$type);
}

export interface Gap extends langium.AstNode {
    readonly $container: Accordion | Button | Card | Grid | IconWidget | LinkWidget | Row | Section | Stack | Tabs | Widget;
    readonly $type: 'Gap';
    value: SpacingValue;
}

export const Gap = {
    $type: 'Gap',
    value: 'value'
} as const;

export function isGap(item: unknown): item is Gap {
    return reflection.isInstance(item, Gap.$type);
}

export interface Grid extends langium.AstNode {
    readonly $container: AccordionItem | Card | Grid | Row | Section | Stack | Tab | ViewDecl;
    readonly $type: 'Grid';
    body: Array<BodyElement>;
    cols: number;
    mods: Array<LayoutMod>;
}

export const Grid = {
    $type: 'Grid',
    body: 'body',
    cols: 'cols',
    mods: 'mods'
} as const;
// ...
export interface IconWidget extends langium.AstNode {
    readonly $container: AccordionItem | Card | Grid | Row | Section | Stack | Tab | ViewDecl;
    readonly $type: 'IconWidget';
    mods: Array<IconModifier>;
    name: string;
}

export const IconWidget = {
    $type: 'IconWidget',
    mods: 'mods',
    name: 'name'
} as const;

export function isIconWidget(item: unknown): item is IconWidget {
    return reflection.isInstance(item, IconWidget.$type);
}

export type IdentName = 'avatar' | 'badge' | 'button' | 'check' | 'color' | 'combobox' | 'currency' | 'date' | 'daterange' | 'datetime' | 'delete' | 'email' | 'exit' | 'expand' | 'file' | 'full' | 'href' | 'icon' | 'json' | 'label' | 'link' | 'long' | 'm' | 'mb' | 'medium' | 'ml' | 'mr' | 'mt' | 'multiselect' | 'mx' | 'my' | 'new' | 'number' | 'on_edit' | 'on_new' | 'on_select' | 'open' | 'p' | 'password' | 'pb' | 'percent' | 'pl' | 'pr' | 'progress' | 'pt' | 'px' | 'py' | 'radio' | 'rangeslider' | 'rating' | 'save' | 'select' | 'short' | 'slider' | 'switch' | 'tags' | 'text' | 'textarea' | 'time' | string;
// ...
export type LayoutMod = Cols | Expand | Gap | Margin | Padding | ScrollFlag;

export const LayoutMod = {
    $type: 'LayoutMod'
} as const;

export function isLayoutMod(item: unknown): item is LayoutMod {
    return reflection.isInstance(item, LayoutMod.$type);
}

export interface LinkWidget extends langium.AstNode {
    readonly $container: AccordionItem | Card | Grid | Row | Section | Stack | Tab | ViewDecl;
    readonly $type: 'LinkWidget';
    caption: string;
    href: string;
    mods: Array<LayoutMod>;
}

export const LinkWidget = {
    $type: 'LinkWidget',
    caption: 'caption',
    href: 'href',
    mods: 'mods'
} as const;

export function isLinkWidget(item: unknown): item is LinkWidget {
    return reflection.isInstance(item, LinkWidget.$type);
}

export interface Margin extends langium.AstNode {
    readonly $container: Accordion | Button | Card | Grid | IconWidget | LinkWidget | Row | Section | Stack | Tabs | Widget;
    readonly $type: 'Margin';
    prop: MarginProp;
    value: SpacingValue;
}

export const Margin = {
    $type: 'Margin',
    prop: 'prop',
    value: 'value'
} as const;

export function isMargin(item: unknown): item is Margin {
    return reflection.isInstance(item, Margin.$type);
}

export type MarginProp = 'm' | 'mb' | 'ml' | 'mr' | 'mt' | 'mx' | 'my';
// ...
export interface Padding extends langium.AstNode {
    readonly $container: Accordion | Button | Card | Grid | IconWidget | LinkWidget | Row | Section | Stack | Tabs | Widget;
    readonly $type: 'Padding';
    prop: PaddingProp;
    value: SpacingValue;
}

export const Padding = {
    $type: 'Padding',
    prop: 'prop',
    value: 'value'
} as const;

export function isPadding(item: unknown): item is Padding {
    return reflection.isInstance(item, Padding.$type);
}

export type PaddingProp = 'p' | 'pb' | 'pl' | 'pr' | 'pt' | 'px' | 'py';
// ...
export interface Row extends langium.AstNode {
    readonly $container: AccordionItem | Card | Grid | Row | Section | Stack | Tab | ViewDecl;
    readonly $type: 'Row';
    body: Array<BodyElement>;
    mods: Array<LayoutMod>;
}

export const Row = {
    $type: 'Row',
    body: 'body',
    mods: 'mods'
} as const;

export function isRow(item: unknown): item is Row {
    return reflection.isInstance(item, Row.$type);
}

export interface ScrollFlag extends langium.AstNode {
    readonly $container: Accordion | Button | Card | Grid | IconWidget | LinkWidget | Row | Section | Stack | Tabs | Widget;
    readonly $type: 'ScrollFlag';
    scroll: BoolLit;
}

export const ScrollFlag = {
    $type: 'ScrollFlag',
    scroll: 'scroll'
} as const;

export function isScrollFlag(item: unknown): item is ScrollFlag {
    return reflection.isInstance(item, ScrollFlag.$type);
}

export interface Section extends langium.AstNode {
    readonly $container: AccordionItem | Card | Grid | Row | Section | Stack | Tab | ViewDecl;
    readonly $type: 'Section';
    body: Array<BodyElement>;
    caption?: string;
    mods: Array<LayoutMod>;
}

export const Section = {
    $type: 'Section',
    body: 'body',
    caption: 'caption',
    mods: 'mods'
} as const;
// ...
export interface SpacingScale extends langium.AstNode {
    readonly $container: Gap | Margin | Padding;
    readonly $type: 'SpacingScale';
    value: number;
}

export const SpacingScale = {
    $type: 'SpacingScale',
    value: 'value'
} as const;

export function isSpacingScale(item: unknown): item is SpacingScale {
    return reflection.isInstance(item, SpacingScale.$type);
}

export interface SpacingToken extends langium.AstNode {
    readonly $container: Gap | Margin | Padding;
    readonly $type: 'SpacingToken';
    token: SpacingTokenName;
}

export const SpacingToken = {
    $type: 'SpacingToken',
    token: 'token'
} as const;

export function isSpacingToken(item: unknown): item is SpacingToken {
    return reflection.isInstance(item, SpacingToken.$type);
}

export type SpacingTokenName = 'lg' | 'md' | 'sm' | 'xl' | 'xs';

export function isSpacingTokenName(item: unknown): item is SpacingTokenName {
    return item === 'xs' || item === 'sm' || item === 'md' || item === 'lg' || item === 'xl';
}

export type SpacingValue = SpacingScale | SpacingToken;

export const SpacingValue = {
    $type: 'SpacingValue'
} as const;

export function isSpacingValue(item: unknown): item is SpacingValue {
    return reflection.isInstance(item, SpacingValue.$type);
}

export interface Stack extends langium.AstNode {
    readonly $container: AccordionItem | Card | Grid | Row | Section | Stack | Tab | ViewDecl;
    readonly $type: 'Stack';
    body: Array<BodyElement>;
    mods: Array<LayoutMod>;
}

export const Stack = {
    $type: 'Stack',
    body: 'body',
    mods: 'mods'
} as const;
// ...
export interface Tab extends langium.AstNode {
    readonly $container: Tabs;
    readonly $type: 'Tab';
    body: Array<BodyElement>;
    caption: string;
}

export const Tab = {
    $type: 'Tab',
    body: 'body',
    caption: 'caption'
} as const;

export function isTab(item: unknown): item is Tab {
    return reflection.isInstance(item, Tab.$type);
}

export interface Table extends langium.AstNode {
    readonly $container: AccordionItem | Card | Grid | Row | Section | Stack | Tab | ViewDecl;
    readonly $type: 'Table';
    actions: Array<ViewAction>;
    columns: Array<Column>;
    rowSource: FieldRef;
}

export const Table = {
    $type: 'Table',
    actions: 'actions',
    columns: 'columns',
    rowSource: 'rowSource'
} as const;

export function isTable(item: unknown): item is Table {
    return reflection.isInstance(item, Table.$type);
}

export interface Tabs extends langium.AstNode {
    readonly $container: AccordionItem | Card | Grid | Row | Section | Stack | Tab | ViewDecl;
    readonly $type: 'Tabs';
    mods: Array<LayoutMod>;
    tabs: Array<Tab>;
}

export const Tabs = {
    $type: 'Tabs',
    mods: 'mods',
    tabs: 'tabs'
} as const;
// ...
export interface ToolbarNew extends langium.AstNode {
    readonly $container: ViewDecl;
    readonly $type: 'ToolbarNew';
    caption: string;
    navMods: Array<NavModifier>;
    target: string;
}

export const ToolbarNew = {
    $type: 'ToolbarNew',
    caption: 'caption',
    navMods: 'navMods',
    target: 'target'
} as const;
// ...
export interface ViewAction extends langium.AstNode {
    readonly $container: Table;
    readonly $type: 'ViewAction';
    event: ViewActionEvent;
    navMods: Array<NavModifier>;
    target: string;
}

export const ViewAction = {
    $type: 'ViewAction',
    event: 'event',
    navMods: 'navMods',
    target: 'target'
} as const;

export function isViewAction(item: unknown): item is ViewAction {
    return reflection.isInstance(item, ViewAction.$type);
}

export type ViewActionEvent = 'on_edit' | 'on_new' | 'on_select';

export function isViewActionEvent(item: unknown): item is ViewActionEvent {
    return item === 'on_select' || item === 'on_new' || item === 'on_edit';
}

export interface ViewDecl extends langium.AstNode {
    readonly $container: ViewModel;
    readonly $type: 'ViewDecl';
    autoRefresh?: AutoRefresh;
    body: Array<BodyElement>;
    default: boolean;
    domains: Array<DomainBinding>;
    name: string;
    title: string;
    toolbar: Array<ToolbarItem>;
}

export const ViewDecl = {
    $type: 'ViewDecl',
    autoRefresh: 'autoRefresh',
    body: 'body',
    default: 'default',
    domains: 'domains',
    name: 'name',
    title: 'title',
    toolbar: 'toolbar'
} as const;
// ...
export interface Widget extends langium.AstNode {
    readonly $container: AccordionItem | Card | Grid | Row | Section | Stack | Tab | ViewDecl;
    readonly $type: 'Widget';
    bind: FieldRef;
    caption?: string;
    kind: WidgetKind;
    mods: Array<WidgetModifier>;
}

export const Widget = {
    $type: 'Widget',
    bind: 'bind',
    caption: 'caption',
    kind: 'kind',
    mods: 'mods'
} as const;
// ...
export interface WidgetFormat extends langium.AstNode {
    readonly $container: Widget;
    readonly $type: 'WidgetFormat';
    fmt: FormatToken;
}

export const WidgetFormat = {
    $type: 'WidgetFormat',
    fmt: 'fmt'
} as const;

export function isWidgetFormat(item: unknown): item is WidgetFormat {
    return reflection.isInstance(item, WidgetFormat.$type);
}

export type WidgetKind = 'avatar' | 'badge' | 'check' | 'color' | 'combobox' | 'date' | 'daterange' | 'email' | 'file' | 'multiselect' | 'number' | 'password' | 'progress' | 'radio' | 'rangeslider' | 'rating' | 'select' | 'slider' | 'switch' | 'text' | 'textarea' | 'time';
// ...
export type WidgetModifier = LayoutMod | WidgetExpand | WidgetFocus | WidgetFormat | WidgetMax | WidgetMin | WidgetPlaceholder | WidgetReadOnly | WidgetStep;

export const WidgetModifier = {
    $type: 'WidgetModifier'
} as const;